requirements.txt
numpy==1.19.5
requests==2.23.0
pickleshare==0.7.5
joblib==1.0.1
pathlib==1.0.1
fastapi==0.68.1
pydantic==1.8.2
portpicker==1.3.9
httplib2==0.17.4
ipykernel==4.10.1
jinja2==2.11.3
keras==2.6.0
astunparse==1.6.3
torchsummary==1.5.1
idna==2.10
werkzeug==1.0.1
uvicorn==0.13.1
ptyprocess==0.7.0
pyparsing==2.4.7
decorator==4.4.2
tqdm==4.62.0
urllib3==1.24.3
gast==0.4.0
cycler==0.10.0
bottleneck==1.3.2
termcolor==1.1.0
pyngrok==5.1.0
ipywidgets==7.6.3
pandas==1.1.5
torchvision
flask==1.1.4
sklearn==0.0
colabcode
python-multipart
torch==1.9.0+cpu torchvision -f https://download.pytorch.org/whl/torch_stable.html
